Correction for angled cameras
Originally, the distortion normal map was stored purely in
screen-space. This caused issues when the camera was at angles relative
to particles, as the relative orientation of the UV offset would change
due to camera rotation.

Now, the UVs are calculated in world-space, and the camera transforms
the distortion map back into camera space before applying the offset.